+++ etc/sql/sql.sql
@@ -158,6 +158,10 @@ CREATE TABLE files (
     owner_id STRING,
     storage_type STRING,
     content_type STRING,
+    dimensions OBJECT(STRICT) AS (
+        width INTEGER,
+        height INTEGER
+    ),
 
     INDEX original_name_ft
       USING FULLTEXT(original_name)